feature: Add RecommendPlaceChats
