+++ feature/characterchat/src/main/java/com/teamoffroad/characterchat/presentation/component/CharacterChatTimeLabel.kt
@@ -4,6 +4,8 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import com.teamoffroad.characterchat.presentation.model.TimeType
 import com.teamoffroad.core.designsystem.theme.Main3
@@ -13,12 +15,14 @@ import java.util.Locale
 @Composable
 fun TimeLabel(
     time: Triple<TimeType, Int, Int>,
+    textStyle: TextStyle = OffroadTheme.typography.textContents,
+    textColor: Color = Main3,
     modifier: Modifier = Modifier,
 ) {
     Text(
         text = "${time.first.krName} ${time.second}:${String.format(Locale.getDefault(), "%02d", time.third)}",
-        color = Main3,
-        style = OffroadTheme.typography.textContents,
+        color = textColor,
+        style = textStyle,
         modifier = modifier
             .padding(horizontal = 6.dp),
     )
